video_subtitle_system/pyproject.toml
[project]
name = "video-subtitle-system"
version = "0.1.0"
description = "离线优先的视频字幕自动化提取系统"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.0.0",
    "yt-dlp>=2024.0.0",
    "imagehash>=4.3.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "httpx>=0.27.0",
    "aiomysql>=0.2.0",
    "redis>=5.0.0",
    "structlog>=24.0.0",
    "PyYAML>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
